moved utils/metadeploy.py to core/metadeploy.py I did this due to some feedback on a code review.
